Rename and Update event display and header components
